+++ docs/source/conf.py
@@ -926,10 +926,22 @@ coverage_ignore_functions = [
     "record_shapeenv_event",
     "replay_shape_env_events",
     "shape_env_check_state_equal",
+    # torch.fx.experimental.sym_node
+    "ceil_impl",
+    "floor_ceil_helper",
+    "floor_impl",
+    "method_to_operator",
+    "sympy_is_channels_last_contiguous_2d",
+    "sympy_is_channels_last_contiguous_3d",
+    "sympy_is_channels_last_strides_2d",
+    "sympy_is_channels_last_strides_3d",
+    "sympy_is_channels_last_strides_generic",
+    "sympy_is_contiguous",
+    "sympy_is_contiguous_generic",
+    "sym_sqrt",
     # torch.fx.experimental.symbolic_shapes
     "bind_symbols",
     "cast_symbool_to_symint_guardless",
-    "ceil_impl",
     "constrain_range",
     "constrain_unify",
     "create_contiguous",
@@ -940,8 +952,6 @@ coverage_ignore_functions = [
     "eval_is_non_overlapping_and_dense",
     "expect_true",
     "find_symbol_binding_fx_nodes",
-    "floor_ceil_helper",
-    "floor_impl",
     "free_symbols",
     "free_unbacked_symbols",
     "fx_placeholder_targets",
@@ -963,18 +973,9 @@ coverage_ignore_functions = [
     "is_non_overlapping_and_dense_indicator",
     "is_symbol_binding_fx_node",
     "is_symbolic",
-    "method_to_operator",
     "parallel_and",
     "parallel_or",
-    "sym_sqrt",
     "sym_ite",
-    "sympy_is_channels_last_contiguous_2d",
-    "sympy_is_channels_last_contiguous_3d",
-    "sympy_is_channels_last_strides_2d",
-    "sympy_is_channels_last_strides_3d",
-    "sympy_is_channels_last_strides_generic",
-    "sympy_is_contiguous",
-    "sympy_is_contiguous_generic",
     "tensor_has_hints",
     "to_node",
     "wrap_node",
@@ -2851,6 +2852,8 @@ coverage_ignore_classes = [
     "RewritingTracer",
     # torch.fx.experimental.schema_type_annotation
     "AnnotateTypesWithSchema",
+    # torch.fx.experimental.sym_node
+    "SymNode",
     # torch.fx.experimental.symbolic_shapes
     "Constraint",
     "ConstraintViolationError",
@@ -2859,17 +2862,13 @@ coverage_ignore_classes = [
     "DynamicDimConstraintPrinter",
     "EqualityConstraint",
     "GuardOnDataDependentSymNode",
-    "IsNonOverlappingAndDenseIndicator",
     "LoggingShapeGuardPrinter",
-    "Pow",
     "RelaxedUnspecConstraint",
     "RuntimeAssert",
     "ShapeEnv",
     "ShapeGuardPrinter",
     "StrictMinMaxConstraint",
     "SymDispatchMode",
-    "SymNode",
-    "TrueDiv",
     # torch.fx.experimental.unification.match
     "Dispatcher",
     "VarDispatcher",
