+++ test/allowlist_for_publicAPI.json
@@ -1971,6 +1971,47 @@
     "sym_function_mode",
     "set_sym_function_mode"
   ],
+  "torch.fx.experimental.sym_node": [
+    "SymNode",
+    "method_to_operator",
+    "magic_methods",
+    "ceil_impl",
+    "floor_impl",
+    "floor_ceil_helper",
+    "is_channels_last_contiguous_2d",
+    "is_channels_last_contiguous_3d",
+    "is_channels_last_strides_2d",
+    "is_channels_last_strides_3d",
+    "is_non_overlapping_and_dense_indicator",
+    "sympy_is_channels_last_contiguous_2d",
+    "sympy_is_channels_last_contiguous_3d",
+    "sympy_is_channels_last_strides_2d",
+    "sympy_is_channels_last_strides_3d",
+    "sympy_is_channels_last_strides_generic",
+    "is_contiguous",
+    "sympy_is_contiguous",
+    "sympy_is_contiguous_generic",
+    "sympy_abs",
+    "sympy_and",
+    "sympy_eq",
+    "sympy_floordiv",
+    "sympy_ge",
+    "sympy_gt",
+    "sympy_ite",
+    "sympy_le",
+    "sympy_lshift",
+    "sympy_lt",
+    "sympy_max",
+    "sympy_min",
+    "sympy_mod",
+    "sympy_ne",
+    "sympy_or",
+    "sympy_pow",
+    "sympy_rshift",
+    "sympy_sqrt",
+    "sympy_truediv",
+    "sym_sqrt"
+  ],
   "torch.fx.experimental.symbolic_shapes": [
     "Constraint",
     "ConstraintViolationError",
@@ -1979,18 +2020,13 @@
     "DynamicDimConstraintPrinter",
     "EqualityConstraint",
     "GuardOnDataDependentSymNode",
-    "IsNonOverlappingAndDenseIndicator",
     "LoggingShapeGuardPrinter",
-    "Pow",
     "RelaxedUnspecConstraint",
     "RuntimeAssert",
     "ShapeGuardPrinter",
     "StrictMinMaxConstraint",
-    "SymNode",
-    "TrueDiv",
     "bind_symbols",
     "cast_symbool_to_symint_guardless",
-    "ceil_impl",
     "constrain_range",
     "constrain_unify",
     "definitely_false",
@@ -2000,31 +2036,15 @@
     "eval_is_non_overlapping_and_dense",
     "expect_true",
     "find_symbol_binding_fx_nodes",
-    "floor_ceil_helper",
-    "floor_impl",
     "free_unbacked_symbols",
     "fx_placeholder_targets",
     "fx_placeholder_vals",
     "guard_bool",
     "has_hint",
-    "is_channels_last_contiguous_2d",
-    "is_channels_last_contiguous_3d",
-    "is_channels_last_strides_2d",
-    "is_channels_last_strides_3d",
-    "is_contiguous",
-    "is_non_overlapping_and_dense_indicator",
     "is_symbolic",
     "parallel_and",
     "parallel_or",
     "safe_expand",
-    "sym_sqrt",
-    "sympy_is_channels_last_contiguous_2d",
-    "sympy_is_channels_last_contiguous_3d",
-    "sympy_is_channels_last_strides_2d",
-    "sympy_is_channels_last_strides_3d",
-    "sympy_is_channels_last_strides_generic",
-    "sympy_is_contiguous",
-    "sympy_is_contiguous_generic",
     "tensor_has_hints",
     "to_node",
     "uninteresting_files"
