+++ torch/fx/experimental/symbolic_shapes.py
@@ -28,6 +28,7 @@ from torch.fx.experimental.recording import (
     replay_shape_env_events,
     shape_env_check_state_equal
 )
+from torch.fx.experimental.sym_node import SymNode
 from torch.fx.experimental.sym_dispatch_mode import set_sym_function_mode, sym_function_mode
 
 # NB: The sym_* functions are used via getattr() and must be imported here.
@@ -43,7 +44,7 @@ from torch import (  # noqa: F401
 )
 from torch._guards import ShapeGuard, Source, TracingContext
 from torch.utils._python_dispatch import is_traceable_wrapper_subclass
-from torch.utils._sympy.functions import FloorDiv, LShift, Mod, RShift
+from torch.utils._sympy.functions import FloorDiv, Mod, IsNonOverlappingAndDenseIndicator
 from torch.utils._sympy.solve import try_solve
 from torch.utils._sympy.value_ranges import bound_sympy, SymPyValueRangeAnalysis, ValueRanges, ValueRangeError
 from torch.utils._sympy.singleton_int import SingletonInt
@@ -68,7 +69,7 @@ aten = torch._ops.ops.aten  # type: ignore[has-type]
 __all__ = [
     "has_symbolic_sizes_strides", "create_contiguous", "ShapeEnv", "is_concrete_int",
     "guard_int", "guard_float", "guard_scalar", "wrap_node",
-    "method_to_operator", "hint_int", "SYMPY_INTERP", "free_symbols", "is_symbol_binding_fx_node",
+    "hint_int", "SYMPY_INTERP", "free_symbols", "is_symbol_binding_fx_node",
     "is_concrete_bool", "SHAPEENV_EVENT_KEY", "CURRENT_NODE_KEY",
 ]
 
@@ -309,6 +310,7 @@ def _constrain_symbol_range(shape_env, s: sympy.Symbol, compiler_min: int, compi
     else:
         shape_env.runtime_var_to_range[s] = ValueRanges(runtime_min, runtime_max)
 
+
 def _advise_is_size(a):
     """
     Don't use this directly; use torch._check_is_size instead.
@@ -341,7 +343,12 @@ def _advise_is_size(a):
     # that hints have to be consistent with static analysis!  If you somehow
     # have an unbounded SymInt that later constrains to 1, this will be
     # inconsistent with the range
-    if isinstance(a, SymInt) and isinstance(a.node, SymNode) and not a.node.has_hint() and isinstance(a.node.expr, sympy.Symbol):
+    if (
+        isinstance(a, SymInt)
+        and isinstance(a.node, SymNode)
+        and not a.node.has_hint()
+        and isinstance(a.node.expr, sympy.Symbol)
+    ):
         _constrain_range_for_size(a)
 
 @record_shapeenv_event()
@@ -542,12 +549,6 @@ def guard_float(a):
     assert isinstance(a, float), a
     return a
 
-# Drop in replacement for math.sqrt
-def sym_sqrt(a):
-    if hasattr(a, '__sym_sqrt__'):
-        return a.__sym_sqrt__()
-    return math.sqrt(a)
-
 def to_node(self, num):
     if isinstance(num, SymTypes):
         return num.node
@@ -724,365 +725,10 @@ class EqualityConstraint(Constraint):
     def is_equal(self, source1, source2):
         return self._find(source1) == self._find(source2)
 
-
-# TODO: An incomplete list
-# 1. Set variables to be equal when we do equality
-# 2. Specialize on 0/1 when we do subtraction
-class SymNode:
-    """
-    This is a type erased SymInt/SymFloat which we use to do actual operations.
-    End users don't touch this.  Magic methods are NOT defined on this object.
-    """
-    def __init__(self, expr, shape_env, pytype, hint: Optional[Union[int, float]], constant=None, fx_node=None):
-        self._expr = expr
-        self.shape_env = shape_env
-        self.pytype = pytype
-        # What's the difference between hint and constant?
-        #
-        # - A constant is known to be invariant across invocations of the model;
-        #   it will always be this value.  We only really know this when we
-        #   encounter an honest-to-goodness literal (when wrapping it into
-        #   a SymNode, we set constant.)  Most of the time, constant is None
-        #
-        # - A hint is a *particular* value from the particular run we are
-        #   tracing, but it may vary the next time around.  It's useful to
-        #   keep this around, as if we need a concrete value from a SymNode,
-        #   we will return the hint and guard on the expression that produced
-        #   it giving the same hint next time around.  The hint is not
-        #   guaranteed to be set either: if you have an unbacked SymNode,
-        #   there won't be any hint; it was the result of some tensor-dependent
-        #   computation, but we don't know what it actually is because we
-        #   haven't actually run the tensor computation.
-        #
-        # hint_expr is only set if we don't have a hint.  When it is set, it
-        # contains the expression which contains the unbacked symnodes that,
-        # if constrained, would allow this expression to be hinted again.
-        if hint is None:
-            self._hint_expr = self.expr.xreplace(shape_env.var_to_val)
-            self._hint = None
-            self._update_hint()  # check if the replacement actually was enough
-        else:
-            self._hint_expr = None
-            self._hint = hint
-        self.constant: Optional[Union[int, float, bool]] = constant
-        # Record the FX node of the current node if we are doing translation
-        # validation. They will be used for building the input assertions for
-        # the translation validation problem.
-        self.fx_node = fx_node if _translation_validation_enabled() else None
-
-    def with_shape_env(self, shape_env: "ShapeEnv") -> "SymNode":
-        return SymNode(self._expr, shape_env, self.pytype, self._hint, self.constant, self.fx_node)
-
-    @property
-    def expr(self):
-        return self.shape_env.replace(self._expr)
-
-    # Check if we have replacements hint_expr that would allow us to
-    # simplify it into a hint
-    def _update_hint(self):
-        if self._hint_expr.free_symbols <= self.shape_env.replacements.keys():
-            new_hint = self.shape_env.replace(self._hint_expr)
-            # NB: unification constraints could result in a replacement that
-            # doesn't actually solve the hint!  Check for this.
-            if new_hint.free_symbols:
-                self._hint_expr = new_hint
-                return
-            self._hint = self.pytype(new_hint)
-            self._hint_expr = None
-
-    @property
-    def hint(self):
-        if self._hint is None:
-            self._update_hint()
-        return self._hint
-
-    def has_hint(self):
-        return self._hint is not None
-
-    def require_hint(self):
-        if self._hint is None:
-            self._update_hint()
-            if self._hint is None:
-                raise self.shape_env._make_data_dependent_error(self._hint_expr, self.expr)
-            else:
-                return self._hint
-        else:
-            return self._hint
-
-    def maybe_as_int(self):
-        if self.expr.free_symbols:
-            return None
-        else:
-            return int(self.expr)
-
-    def is_int(self):
-        return self.pytype is int
-
-    def is_float(self):
-        return self.pytype is float
-
-    def is_bool(self):
-        return self.pytype is bool
-
-    def wrap_int(self, num):
-        assert type(num) is int
-        return SymNode(sympy.Integer(num), self.shape_env, int, num, constant=num, fx_node=num)
-
-    def wrap_float(self, num):
-        assert type(num) is float
-        return SymNode(sympy.Float(num), self.shape_env, float, num, constant=num, fx_node=num)
-
-    def wrap_bool(self, num):
-        assert type(num) is bool
-        return SymNode(sympy.true if num else sympy.false, self.shape_env, bool, num, constant=num, fx_node=num)
-
-    def clone(self):
-        return self
-
-    def str(self):
-        return f"{self.expr}"
-
-    def __str__(self):
-        return self.str()
-
-    def __repr__(self):
-        return self.str()
-
-    # These methods call the metaprogrammed methods, they're hand written
-    # here so we get good stack traces
-    def abs(self) -> "SymNode":  # noqa: F811
-        return self._abs()  # type: ignore[attr-defined]
-
-    def add(self, other) -> "SymNode":  # noqa: F811
-        return self._add(other)  # type: ignore[attr-defined]
-
-    def sub(self, other) -> "SymNode":  # noqa: F811
-        return self._sub(other)  # type: ignore[attr-defined]
-
-    def mul(self, other) -> "SymNode":  # noqa: F811
-        return self._mul(other)  # type: ignore[attr-defined]
-
-    def mod(self, other) -> "SymNode":  # noqa: F811
-        return self._mod(other)  # type: ignore[attr-defined]
-
-    def pow(self, other) -> "SymNode":  # noqa: F811
-        return self._pow(other)  # type: ignore[attr-defined]
-
-    def and_(self, other) -> "SymNode":  # noqa: F811
-        return self._and_(other)  # type: ignore[attr-defined]
-
-    def or_(self, other) -> "SymNode":  # noqa: F811
-        return self._or_(other)  # type: ignore[attr-defined]
-
-    def truediv(self, other) -> "SymNode":  # noqa: F811
-        return self._truediv(other)  # type: ignore[attr-defined]
-
-    def floordiv(self, other) -> "SymNode":  # noqa: F811
-        return self._floordiv(other)  # type: ignore[attr-defined]
-
-    def lshift(self, other) -> "SymNode":  # noqa: F811
-        return self._lshift(other)  # type: ignore[attr-defined]
-
-    def rshift(self, other) -> "SymNode":  # noqa: F811
-        return self._rshift(other)  # type: ignore[attr-defined]
-
-    def sym_not(self) -> "SymNode":  # noqa: F811
-        return self._sym_not()  # type: ignore[attr-defined]
-
-    def eq(self, other) -> "SymNode":  # noqa: F811
-        return self._eq(other)  # type: ignore[attr-defined]
-
-    def ne(self, other) -> "SymNode":  # noqa: F811
-        return self._ne(other)  # type: ignore[attr-defined]
-
-    def gt(self, other) -> "SymNode":  # noqa: F811
-        return self._gt(other)  # type: ignore[attr-defined]
-
-    def lt(self, other) -> "SymNode":  # noqa: F811
-        return self._lt(other)  # type: ignore[attr-defined]
-
-    def le(self, other) -> "SymNode":  # noqa: F811
-        return self._le(other)  # type: ignore[attr-defined]
-
-    def ge(self, other) -> "SymNode":  # noqa: F811
-        return self._ge(other)  # type: ignore[attr-defined]
-
-    def floor(self) -> "SymNode":  # noqa: F811
-        return self._floor()  # type: ignore[attr-defined]
-
-    def sym_float(self) -> "SymNode":  # noqa: F811
-        return self._sym_float()  # type: ignore[attr-defined]
-
-    def sym_int(self) -> "SymNode":  # noqa: F811
-        return self._sym_int()  # type: ignore[attr-defined]
-
-    def ceil(self) -> "SymNode":  # noqa: F811
-        return self._ceil()  # type: ignore[attr-defined]
-
-    def neg(self) -> "SymNode":  # noqa: F811
-        return self._neg()  # type: ignore[attr-defined]
-
-    def sym_min(self, other) -> "SymNode":  # noqa: F811
-        return self._sym_min(other)  # type: ignore[attr-defined]
-
-    def sym_max(self, other) -> "SymNode":  # noqa: F811
-        return self._sym_max(other)  # type: ignore[attr-defined]
-
-    def sym_ite(self, then_val, else_val) -> "SymNode":
-        return self._sym_ite(then_val, else_val)
-
-    def sym_sqrt(self) -> "SymNode":  # noqa: F811
-        return self._sym_sqrt()  # type: ignore[attr-defined]
-
-    def is_contiguous(self, sizes, strides) -> "SymNode":  # noqa: F811
-        return self._is_contiguous(sizes, strides)  # type: ignore[attr-defined]
-
-    def is_channels_last_contiguous_2d(self, sizes, strides) -> "SymNode":  # noqa: F811
-        return self._is_channels_last_contiguous_2d(sizes, strides)  # type: ignore[attr-defined]
-
-    def is_channels_last_contiguous_3d(self, sizes, strides) -> "SymNode":  # noqa: F811
-        return self._is_channels_last_contiguous_3d(sizes, strides)  # type: ignore[attr-defined]
-
-    def is_channels_last_strides_2d(self, sizes, strides) -> "SymNode":  # noqa: F811
-        return self._is_channels_last_strides_2d(sizes, strides)  # type: ignore[attr-defined]
-
-    def is_channels_last_strides_3d(self, sizes, strides) -> "SymNode":  # noqa: F811
-        return self._is_channels_last_strides_3d(sizes, strides)  # type: ignore[attr-defined]
-
-    def is_non_overlapping_and_dense_indicator(self, sizes, strides) -> "SymNode":  # noqa: F811
-        return self._is_non_overlapping_and_dense_indicator(sizes, strides)  # type: ignore[attr-defined]
-
-    # Make C++ happy
-    def sym_or(self, other):  # noqa: F811
-        return self.or_(other)
-
-    def sym_and(self, other):  # noqa: F811
-        return self.and_(other)
-
-    def is_non_overlapping_and_dense(self, sizes, strides):
-        return self.is_non_overlapping_and_dense_indicator(sizes, strides).eq(to_node(self, 1))  # type: ignore[attr-defined]
-
-    def int_(self):
-        return self.guard_int("", 0)  # NB: uses Python backtrace
-
-    # You can manually trigger a guard with this function
-    def guard_int(self, file, line):
-        # TODO: use the file/line for some useful diagnostic on why a
-        # guard occurred
-        r = self.shape_env.evaluate_expr(self.expr, self.hint, fx_node=self.fx_node)
-        try:
-            return int(r)
-        except Exception:
-            log.warning("Failed to convert to int: %s", r)
-            raise
-
-    def guard_float(self, file, line):
-        # TODO: use the file/line for some useful diagnostic on why a
-        # guard occurred
-        r = self.shape_env.evaluate_expr(self.expr, self.hint, fx_node=self.fx_node)
-        try:
-            return float(r)
-        except Exception:
-            log.warning("Failed to convert to float: %s", r)
-            raise
-
-    def guard_bool(self, file, line):
-        # TODO: use the file/line for some useful diagnostic on why a
-        # guard occurred
-        r = self.shape_env.evaluate_expr(self.expr, self.hint, fx_node=self.fx_node)
-        try:
-            return bool(r)
-        except Exception:
-            log.warning("Failed to convert to bool: %s", r)
-            raise
-
-    def expect_true(self, file, line):
-        if self.has_hint():
-            # OK to generate guards
-            return self.guard_bool(file, line)
-        # Generate a deferred runtime assert (this might actually end up doing
-        # a regular guard if we can!)
-        # TODO: file/line here is very important, because the assert has been
-        # deferred so you can't backtrace easily
-        return self.shape_env.defer_runtime_assert(self.expr, f"{file}:{line}", fx_node=self.fx_node)
-
-    def expect_size(self, file, line):
-        b = self.ge(self.wrap_int(0))
-        # Generate a deferred runtime assert
-        r = b.expect_true(file, line)
-        # Refine compile time range, but only if it's unbacked.
-        # If you refine range for hinted variables, you can end up making
-        # improper deductions since compile time reasoning may be
-        # incompatible with runtime reasoning.
-        if r and not self.has_hint():
-            _advise_is_size(SymInt(self))
-        return r
-
-    def bool_(self):
-        return self.guard_bool("", 0)
-
-    def is_symbolic(self):
-        return True
-
-    def singleton_int(self):
-        return None
-
-    def is_constant(self):
-        return False
-
 def is_symbolic(val: Union[int, SymInt, float, SymFloat, bool, SymBool]) -> bool:
     if isinstance(val, (int, float, bool)):
         return False
     return val.node.is_symbolic()
-
-# Overloaded to be compatible with regular Python.
-# https://github.com/pytorch/pytorch/issues/90900
-class Pow(sympy.Function):
-    @classmethod
-    def eval(cls, base, exp):
-        if exp.is_zero:
-            return sympy.Integer(1)
-        elif base.is_zero and exp < 0:
-            raise ZeroDivisionError(f"{base} cannot be raised to a negative power")
-        else:
-            return base ** exp
-
-# Overloaded to be compatible with regular Python.
-# https://github.com/pytorch/pytorch/issues/90900
-class TrueDiv(sympy.Function):
-    @classmethod
-    def eval(cls, base, divisor):
-        if divisor.is_zero:
-            raise ZeroDivisionError("division by zero")
-        else:
-            return base / divisor
-
-# TODO: As an indicator, this != 0 implies == 1 (and vice versa).
-# Because we do not have the ability to guard on the stride permutation
-# at the moment, it is hard to make further inferences when this is true,
-# as although we know the tensor is contiguous in *some* layout, we don't
-# know which one (however, you could, for example, make the inference that
-# reshaping this to a 1D tensor can be guard-free.)
-class IsNonOverlappingAndDenseIndicator(sympy.Function):
-    is_integer = True
-
-    @classmethod
-    def eval(cls, *args):
-        assert len(args) % 2 == 0
-        dim = len(args) // 2
-        # TODO: it is possible to make progress evaluating this guard
-        # even if not all of the inputs are known.  For example, a 2D
-        # tensor with non-0/1 sizes but strides (0, 1) is definitely
-        # false, because we know its numel > 1 but it's broadcasted
-        # in dim 0.
-        if all(isinstance(a, sympy.Integer) for a in args):
-            size_args = args[0:dim]
-            stride_args = args[dim:]
-            return eval_is_non_overlapping_and_dense(
-                [int(a) for a in size_args],
-                [int(a) for a in stride_args]
-            )
-        return None
 
 IndicatorTypes = (IsNonOverlappingAndDenseIndicator,)
 
@@ -1097,153 +743,13 @@ def safe_expand(r):
     else:
         return r
 
-# Methods that have a `__foo__` as well as `__rfoo__`
-reflectable_magic_methods = {
-    'add': lambda a, b: a + b,
-    'sub': lambda a, b: a - b,
-    'mul': lambda a, b: a * b,
-    'mod': lambda a, b: Mod(a, b),
-    'pow': lambda a, b: Pow(a, b),
-    'and': lambda a, b: sympy.And(a, b),
-    'or': lambda a, b: sympy.Or(a, b),
-    'truediv': lambda a, b: TrueDiv(a, b),
-    'floordiv': lambda a, b: FloorDiv(a, b),
-    'lshift': lambda a, b: LShift(a, b),
-    'rshift': lambda a, b: RShift(a, b),
-}
-
-
 def error():
     raise AssertionError("shouldn't be hit")
 
 
-def floor_ceil_helper(a, fn):
-    if isinstance(a, sympy.Mul):
-        aa = a.args
-        if len(aa) == 2 and isinstance(aa[0], sympy.Float) and aa[1].is_integer:
-            coef = sympy.Integer(aa[0])
-            if aa[0] == coef:  # structural equality test
-                return coef * aa[1]
-    if isinstance(a, sympy.Float) and a == sympy.Integer(a) or isinstance(a, sympy.Integer):
-        return sympy.Integer(a)
-    return fn(a)
-
-def floor_impl(a):
-    return floor_ceil_helper(a, sympy.floor)
-
-def ceil_impl(a):
-    return floor_ceil_helper(a, sympy.ceiling)
-
-
-magic_methods = {
-    **reflectable_magic_methods,
-    'sym_not': lambda a: ~a,
-    'eq': lambda a, b: sympy.Eq(a, b),
-    'ne': lambda a, b: sympy.Ne(a, b),
-    'gt': lambda a, b: sympy.Gt(a, b),
-    'lt': lambda a, b: sympy.Lt(a, b),
-    'le': lambda a, b: sympy.Le(a, b),
-    'ge': lambda a, b: sympy.Ge(a, b),
-    'floor': floor_impl,
-    'sym_float': lambda a: a,  # Cannot use sympy.Float(a) here, coz it expects python literals
-    'ceil': ceil_impl,
-    'neg': lambda a: -a,
-    'sym_min': lambda a, b: sympy.Min(a, b),
-    'sym_max': lambda a, b: sympy.Max(a, b),
-    'sym_ite': lambda a, t, f: sympy.Piecewise((t, a), (f, True)),
-    'sym_sqrt': lambda a: sympy.sqrt(a),
-    'abs': lambda a: sympy.Abs(a),
-}
-
-sizes_strides_methods = {
-    # TODO: These could also be done with indicators, maybe it is better
-    # for reasoning to do it that way
-    'is_contiguous': lambda sizes, strides: sympy_is_contiguous(sizes, strides),
-    'is_channels_last_contiguous_2d': lambda sizes, strides: sympy_is_channels_last_contiguous_2d(sizes, strides),
-    'is_channels_last_contiguous_3d': lambda sizes, strides: sympy_is_channels_last_contiguous_3d(sizes, strides),
-    'is_channels_last_strides_2d': lambda sizes, strides: sympy_is_channels_last_strides_2d(sizes, strides),
-    'is_channels_last_strides_3d': lambda sizes, strides: sympy_is_channels_last_strides_3d(sizes, strides),
-    'is_non_overlapping_and_dense_indicator': lambda sizes, strides: IsNonOverlappingAndDenseIndicator(*sizes, *strides),
-}
-
-alternate_impl_if_hinted_methods = {
-    "sym_min": builtins.min,
-    "sym_max": builtins.max,
-}
-
-def sympy_is_contiguous_generic(sizes, strides, dim_order):
-    dim = len(sizes)
-
-    if len(dim_order) != dim:
-        return sympy.false
-
-    is_contiguous = sympy.true
-    z = sympy.Integer(1)
-    # Contiguous if the strides make sense (or the dim is size 1)
-    for d in dim_order:
-        is_contiguous &= sympy.Eq(sizes[d], sympy.Integer(1)) | sympy.Eq(strides[d], z)
-        z *= sizes[d]
-    # OR if any size is zero
-    for d in range(dim):
-        is_contiguous |= sympy.Eq(sizes[d], sympy.Integer(0))
-    return is_contiguous
-
-def sympy_is_contiguous(sizes, strides):
-    dim = len(sizes)
-    return sympy_is_contiguous_generic(sizes, strides, list(range(dim - 1, -1, -1)))
-
-# NB: There is a TODO in C++ to allow omitting the batch dim.  If that
-# happens you will need to refactor this
-
-def sympy_is_channels_last_contiguous_2d(sizes, strides):
-    return sympy_is_contiguous_generic(sizes, strides, [1, 3, 2, 0])
-
-def sympy_is_channels_last_contiguous_3d(sizes, strides):
-    return sympy_is_contiguous_generic(sizes, strides, [1, 4, 3, 2, 0])
-
-def sympy_is_channels_last_strides_generic(sizes, strides, dim_order):
-    dim = len(sizes)
-
-    if dim != len(dim_order):
-        return sympy.false
-
-    m = sympy.Integer(0)
-    r = sympy.true
-
-    # special case for trivial C dimension. default to NCHW
-    r &= sympy.Ne(strides[1], 0)
-
-    for d in dim_order:
-        r &= sympy.Ne(sizes[d], 0) & (strides[d] >= m)
-        # Fallback to NCHW as default layout for ambiguous cases
-        # This is the flaw of implicit memory_format from strides.
-        # N111 tensor with identical strides for size 1 dimension;
-        # Two cases could lead us here:
-        # a. N111 contiguous Tensor ([N,1,1,1]@[1,1,1,1])
-        # b. N11W contiguous Tensor sliced on the W-dimension.
-        # ([N,1,1,1]@[W,W,W,W])
-        if d == 0:
-            r &= sympy.Ne(m, strides[1])
-        # This is necessary to:
-        # 1. distinguish the memory_format of N1H1;
-        #     [H, 1, 1, 1] channels_last stride
-        #     [H, H, 1, 1] contiguous stride
-        # 2. permutation of 1C1W:
-        #     [1, C, 1, H]@[HC, H, H, 1] transpose(1, 3)
-        #     [1, H, 1, C]@[HC, 1, H, H] shouldn't be identified as
-        #     channels_last
-        m = strides[d] * sympy.Max(sizes[d], 1)
-
-    return r
-
-def sympy_is_channels_last_strides_2d(sizes, strides):
-    return sympy_is_channels_last_strides_generic(sizes, strides, [1, 3, 2, 0])
-
-def sympy_is_channels_last_strides_3d(sizes, strides):
-    return sympy_is_channels_last_strides_generic(sizes, strides, [1, 4, 3, 2, 0])
-
 # TODO: Deduplicate this with torch/_prims_common/__init__.py
 def eval_is_non_overlapping_and_dense(sizes, strides):
+    from torch.fx.experimental.symbolic_shapes import guard_bool
     return int(guard_bool(_eval_is_non_overlapping_and_dense(sizes, strides)))
 
 def _eval_is_non_overlapping_and_dense(sizes, strides):
@@ -1276,39 +782,6 @@ def _eval_is_non_overlapping_and_dense(sizes, strides):
 
     return True
 
-unary_magic_methods = {
-    'abs',
-    'sym_float',
-    'ceil',
-    'floor',
-    'neg',
-    'sym_sqrt',
-    'sym_not',
-}
-
-# Most methods are only registered on SymInt and SymFloat
-# Some methods are only be registered on SymBool
-only_bool_magic_methods = {"and", "or", "sym_not", "sym_ite"}
-# Methods that are also on SymBool, in addition to on SymInt and SymFloat
-also_bool_magic_methods = {"eq"}
-bool_magic_methods = only_bool_magic_methods | also_bool_magic_methods
-
-magic_methods_on_math = {"ceil", "floor"}
-magic_methods_on_submodule = {"sym_float", "sym_sqrt", "sym_min", "sym_max", "sym_not", "sym_ite"}
-magic_methods_on_operator_with_trailing_underscore = {"and", "or"}
-
-def method_to_operator(method):
-    if method in magic_methods_on_operator_with_trailing_underscore:
-        method_attr = f"{method}_"
-    else:
-        method_attr = method
-    if method in magic_methods_on_submodule:
-        op = getattr(torch.fx.experimental.symbolic_shapes, method_attr)
-    elif method in magic_methods_on_math:
-        op = getattr(math, method_attr)
-    else:
-        op = getattr(operator, method_attr)
-    return op
 
 def cast_symbool_to_symint_guardless(symbool: torch.SymBool) -> torch.SymInt:
     int_sym = sympy.Piecewise((1, symbool.node.expr), (0, True))
@@ -1333,10 +806,6 @@ SYMPY_INTERP = {
     'cast_symbool_to_symint_guardless': cast_symbool_to_symint_guardless,
 }
 
-always_float_magic_methods = {"truediv", "sym_float", "sym_sqrt", "pow"}
-always_int_magic_methods = {"ceil", "floor"}
-always_bool_magic_methods = {"eq", "ne", "gt", "lt", "le", "ge", "and", "or", "sym_not", "is_non_overlapping_and_dense"}
-
 def wrap_node(x):
     # TODO: let C++ also take advantage of this
     if isinstance(x, SymNode) and x.constant is not None:
@@ -1349,288 +818,6 @@ def wrap_node(x):
         return SymBool(x)
     else:
         raise AssertionError(f"unrecognized return type {x}")
-
-def _make_node_magic(method, func):
-    func = lru_cache(256)(func)
-
-    if method in magic_methods_on_operator_with_trailing_underscore:
-        method_attr = f"{method}_"
-    else:
-        method_attr = method
-
-    def binary_magic_impl(self, other):
-        op = method_to_operator(method)
-
-        out_hint = None
-        if self.hint is not None and other.hint is not None:
-            out_hint = op(self.hint, other.hint)
-
-        alternate_impl = alternate_impl_if_hinted_methods.get(method)
-        if alternate_impl and out_hint is not None:
-            return to_node(self, alternate_impl(wrap_node(self), wrap_node(other)))
-
-        if sym_function_mode():
-            return to_node(self, _handle_sym_dispatch(op, (wrap_node(self), wrap_node(other)), {}))
-        assert isinstance(other, SymNode)
-        # TODO: consider constant prop here
-        try:
-            out = func(self.expr, other.expr)
-        except Exception:
-            log.warning("failed to eval %s(%s, %s)", method, self.expr, other.expr)
-            raise
-        out = safe_expand(out)
-        pytype: Type
-        # This is not strictly correct. In Python, a**b may return complex when
-        # a < 0 and b is a float: (-1)**2.1. Same for sympy.sqrt(-3.14). This
-        # returns a float while both arguments are ints: 2**(-1). Also, max and
-        # min do not type promote. To avoid having data-dependent control flow
-        # here, we just set the type to float if one of the args is a float. In
-        # case of a type mismatch, we assume that it will be detected during
-        # evaluation.
-        if method in always_float_magic_methods:
-            pytype = float
-        elif method in always_bool_magic_methods:
-            pytype = bool
-        elif self.pytype is float or other.pytype is float:
-            pytype = float
-        else:
-            pytype = self.pytype
-
-        # Create a FX node that corresponds to the operation being applied to
-        # this node.
-        fx_node, _ = self.shape_env.create_fx_call_function(op, (self.fx_node, other.fx_node))
-        return SymNode(out, self.shape_env, pytype, out_hint, fx_node=fx_node)
-
-    def unary_magic_impl(self):
-        op = method_to_operator(method)
-        if sym_function_mode():
-            return to_node(self, _handle_sym_dispatch(op, (wrap_node(self),), {}))
-        # TODO: consider constant prop here
-        expr = self.expr
-        if method == "floor" or method == "ceiling":
-            expr = self.shape_env._simplify_floor_div(expr)
-
-        try:
-            out = func(expr)
-        except Exception:
-            log.warning("failed to eval %s(%s)", method, expr)
-            raise
-
-        out_hint = None
-        if self.hint is not None:
-            out_hint = op(self.hint)
-        out = safe_expand(out)
-        pytype: Type
-        if method in always_int_magic_methods:
-            pytype = int
-        elif method in always_float_magic_methods:
-            pytype = float
-        else:
-            pytype = self.pytype
-
-        fx_node, _ = self.shape_env.create_fx_call_function(op, (self.fx_node,))
-        return SymNode(out, self.shape_env, pytype, out_hint, fx_node=fx_node)
-
-    if method in unary_magic_methods:
-        setattr(SymNode, f"_{method_attr}", unary_magic_impl)
-    elif method == "sym_ite":
-
-        def sym_ite_impl(pred_node, then_node, else_node):
-            out_hint = then_node.hint if pred_node.hint else else_node.hint
-            if sym_function_mode():
-                return to_node(
-                    pred_node,
-                    _handle_sym_dispatch(
-                        sym_ite,
-                        (wrap_node(pred_node), wrap_node(then_node), wrap_node(else_node)), {}
-                    )
-                )
-
-            try:
-                out = func(pred_node.expr, then_node.expr, else_node.expr)
-            except Exception:
-                log.warning("failed to eval %s(%s, %s, %s)", method, pred_node.expr, then_node.expr, else_node.expr)
-                raise
-
-            out = safe_expand(out)
-            fx_node, _ = pred_node.shape_env.create_fx_call_function(
-                sym_ite,
-                (
-                    pred_node.fx_node,
-                    then_node.fx_node,
-                    else_node.fx_node
-                )
-            )
-            return SymNode(out, pred_node.shape_env, then_node.pytype, out_hint, fx_node=fx_node)
-        setattr(SymNode, f"_{method_attr}", sym_ite_impl)
-    else:
-        setattr(SymNode, f"_{method_attr}", binary_magic_impl)
-
-def _make_node_sizes_strides(method, func):
-    # NB: don't LRU cache, lots of arguments
-
-    def sizes_strides_impl(self, sizes, strides):
-        op = getattr(sys.modules[__name__], method)
-        if sym_function_mode():
-            return to_node(
-                self,
-                _handle_sym_dispatch(
-                    op,
-                    ([wrap_node(s) for s in sizes], [wrap_node(s) for s in strides]),
-                    {}
-                )
-            )
-        size_exprs = [s.expr for s in sizes]
-        stride_exprs = [s.expr for s in strides]
-        try:
-            out = func(size_exprs, stride_exprs)
-        except Exception:
-            log.warning("failed to eval %s(%s, %s)", method, size_exprs, stride_exprs)
-            raise
-        # bool is never expandable
-
-        size_hints = []
-        out_hint = None
-        for s in sizes:
-            if s.hint is None:
-                break
-            size_hints.append(s.hint)
-        else:
-            stride_hints = []
-            for s in strides:
-                if s.hint is None:
-                    break
-                stride_hints.append(s.hint)
-            else:
-                out_hint = op(size_hints, stride_hints)
-
-        # NB: This is the indicator function, not the actual bool!
-        pytype: Type
-        if method.endswith("_indicator"):
-            pytype = int
-        else:
-            pytype = bool
-        return SymNode(out, self.shape_env, pytype, out_hint)
-
-    setattr(SymNode, f"_{method}", sizes_strides_impl)
-
-    # TODO: This is technically hotpath, but in the ideal end state
-    # guards on this will resolve at a higher level so you never
-    # spend time in this code
-    def sizes_strides_user(sizes, strides):
-        for a in itertools.chain(sizes, strides):
-            if isinstance(a, SymInt):
-                return wrap_node(getattr(a.node, method)(
-                    [to_node(a.node, b) for b in sizes],
-                    [to_node(a.node, b) for b in strides],
-                ))
-        if method == "is_non_overlapping_and_dense_indicator":
-            return eval_is_non_overlapping_and_dense(sizes, strides)
-        else:
-            # TODO: this is an awful implementation
-            return bool(func(
-                [sympy.sympify(a) for a in sizes],
-                [sympy.sympify(a) for a in strides],
-            ))
-
-    # Skip for is_non_overlapping_and_dense_indicator
-    if not hasattr(sys.modules[__name__], method):
-        setattr(sys.modules[__name__], method, sizes_strides_user)
-
-for method, func in magic_methods.items():
-    _make_node_magic(method, func)
-
-for method, func in sizes_strides_methods.items():
-    _make_node_sizes_strides(method, func)
-
-def _make_user_magic(method, user_type):
-    # User magic takes care of wrapping the other operand into a node,
-    # so that our internal logic can assume everything is nodes
-
-    if method in magic_methods_on_operator_with_trailing_underscore:
-        method_attr = f"{method}_"
-    else:
-        method_attr = method
-
-    def get_constant(x: Union[SymInt, int, SymFloat, float, SymBool, bool]):
-        if isinstance(x, (int, float, bool)):
-            return x
-        if isinstance(x, SymBool):
-            return x.node.guard_bool("", 0)
-        raise AssertionError("expect to be called with constant SymBools")
-
-    def is_constant(x):
-        if isinstance(x, (int, float, bool)):
-            return True
-        if isinstance(x, (SymInt, SymFloat, SymBool)):
-            return x.node.is_constant()
-        return False
-
-    # Before and after performing the operation, check if any operands are constant.
-    # If so, extract out the constant values first. If `self` itself is a
-    # constant, then "redispatch" by calling back into the operator. Sometimes
-    # this means that operations involving SymBool return plain bools.
-    # Alternatively, we could also rewrap into constant Symbool (i.e. by
-    # implementing wrap_bool in ConstantSymNodeImpl), but we're not doing that
-    # today for no particular reason.
-    def unary_magic_impl(self):
-        if is_constant(self):
-            return (method_to_operator(method))(get_constant(self))
-        return wrap_node(getattr(self.node, method_attr)())
-
-    def binary_magic_impl(self, other):
-        if is_constant(self):
-            return (method_to_operator(method))(get_constant(self), other)
-        if is_constant(other):
-            other = get_constant(other)
-        other_node = to_node(self.node, other)
-        if other_node is NotImplemented:
-            return NotImplemented
-        ret = wrap_node(getattr(self.node, method_attr)(other_node))
-        return get_constant(ret) if is_constant(ret) else ret
-
-    def rbinary_magic_impl(self, other):
-        if is_constant(self):
-            return (method_to_operator(method))(get_constant(self), other)
-        if is_constant(other):
-            other = get_constant(other)
-        other_node = to_node(self.node, other)
-        if other_node is NotImplemented:
-            return NotImplemented
-        ret = wrap_node(getattr(other_node, method_attr)(self.node))
-        return get_constant(ret) if is_constant(ret) else ret
-
-    if method in unary_magic_methods:
-        setattr(user_type, f"__{method}__", unary_magic_impl)
-    elif method == "sym_ite":
-
-        def sym_ite_magic_impl(pred, then_val, else_val):
-            pred_node = pred.node
-            then_node = to_node(pred_node, then_val)
-            else_node = to_node(pred_node, else_val)
-            if then_node is NotImplemented or else_node is NotImplemented:
-                return NotImplemented
-            assert isinstance(then_node, SymNode) and isinstance(else_node, SymNode) and then_node.pytype == else_node.pytype
-            ret = wrap_node(getattr(pred.node, method_attr)(then_node, else_node))
-            return get_constant(ret) if ret.node.is_constant() else ret
-
-        setattr(user_type, f"__{method}__", sym_ite_magic_impl)
-    else:
-        setattr(user_type, f"__{method}__", binary_magic_impl)
-        if method in reflectable_magic_methods:
-            setattr(user_type, f"__r{method}__", rbinary_magic_impl)
-
-for method, func in magic_methods.items():
-    if method in only_bool_magic_methods:
-        _make_user_magic(method, SymBool)
-        continue
-    if method in also_bool_magic_methods:
-        _make_user_magic(method, SymBool)
-    _make_user_magic(method, SymInt)
-    _make_user_magic(method, SymFloat)
-
-del method
-del func
 
 
 def _translation_validation_enabled() -> bool:
